Update element2D create methods for environment engine
