+++ Environment_Engine/Create/NewInternalElement2D.cs
@@ -1,6 +1,6 @@
 /*
  * This file is part of the Buildings and Habitats object Model (BHoM)
- * Copyright (c) 2015 - 2018, the respective contributors. All rights reserved.
+ * Copyright (c) 2015 - 2019, the respective contributors. All rights reserved.
  *
  * Each contributor holds copyright over their respective contributions.
  * The project versioning (Git) records all such contribution source information.
@@ -22,7 +22,6 @@
 
 using BH.oM.Environment.Elements;
 using BH.oM.Geometry;
-using BH.Engine.Geometry;
 
 namespace BH.Engine.Environment
 {
@@ -32,16 +31,9 @@ namespace BH.Engine.Environment
         /**** Public Methods                            ****/
         /***************************************************/
 
-        public static IElement1D NewElement1D(this Opening opening, ICurve curve)
+        public static IElement2D NewInternalElement2D(this Panel panel)
         {
-            return curve.IClone();
-        }
-
-        /***************************************************/
-
-        public static IElement1D NewElement1D(this Panel panel, ICurve curve)
-        {
-            return curve.IClone();
+            return new Opening();
         }
 
         /***************************************************/
